Conflicts:

	other-licenses/branding/sunbird/Makefile.in

+++ calendar/providers/storage/calStorageCalendar.js
@@ -214,9 +214,11 @@ calStorageCalendar.prototype = {
                         if (this.mDB.tableExists("cal_events")) { // check again (with lock)
                             // take over data and drop from storage.sdb tables:
                             for (let table in getSqlTable(DB_SCHEMA_VERSION)) {
-                                this.mDB.executeSimpleSQL("CREATE TABLE local_sqlite." +  table +
-                                                          " AS SELECT * FROM " + table +
-                                                          "; DROP TABLE IF EXISTS " +  table);
+                                if (table.substr(0, 4) != "idx_") {
+                                    this.mDB.executeSimpleSQL("CREATE TABLE local_sqlite." +  table +
+                                                              " AS SELECT * FROM " + table +
+                                                              "; DROP TABLE IF EXISTS " +  table);
+                                }
                             }
                             this.mDB.commitTransaction();
                         } else { // migration done in the meantime
@@ -714,12 +716,6 @@ calStorageCalendar.prototype = {
 
         upgradeDB(this.mDB);
 
-        // (Conditionally) add index
-        this.mDB.executeSimpleSQL(
-            "CREATE INDEX IF NOT EXISTS " + 
-            "idx_cal_properies_item_id ON cal_properties(cal_id, item_id);"
-            );
-
         this.mSelectEvent = createStatement (
             this.mDB,
             "SELECT * FROM cal_events " +
@@ -859,16 +855,18 @@ calStorageCalendar.prototype = {
 
         this.mSelectPropertiesForItem = createStatement(
             this.mDB,
-            "SELECT * FROM cal_properties " +
-            "WHERE item_id = :item_id AND recurrence_id IS NULL"
+            "SELECT * FROM cal_properties" +
+            " WHERE item_id = :item_id" +
+            "   AND cal_id = " + this.mCalId +
+            "   AND recurrence_id IS NULL"
             );
 
         this.mSelectPropertiesForItemWithRecurrenceId = createStatement(
             this.mDB,
             "SELECT * FROM cal_properties " +
             "WHERE item_id = :item_id AND cal_id = " + this.mCalId +
-            " AND recurrence_id = :recurrence_id" +
-            " AND recurrence_id_tz = :recurrence_id_tz"
+            "  AND recurrence_id = :recurrence_id" +
+            "  AND recurrence_id_tz = :recurrence_id_tz"
             );
 
         this.mSelectRecurrenceForItem = createStatement(
@@ -925,7 +923,7 @@ calStorageCalendar.prototype = {
 
         this.mSelectAlarmsForItemWithRecurrenceId = createStatement(
             this.mDB,
-            "SELECT * FROM cal_alarms" +
+            "SELECT icalString FROM cal_alarms" +
             " WHERE item_id = :item_id AND cal_id = " + this.mCalId +
             " AND recurrence_id = :recurrence_id" +
             " AND recurrence_id_tz = :recurrence_id_tz"
